Server/requirements.txt
fastapi
uvicorn
langgraph
langchain
pymongo
python-dotenv
requests
beautifulsoup4
openai
selenium
webdriver-manager